No console window or debugging on release mode

+++ OpenGL/src/Application.cpp
@@ -90,7 +90,7 @@ void Application::imGuiRender()
 	GUI::renderNewObjectMenu();
 	GUI::renderToolbar();
 
-	ImGui::SetNextWindowPos(ImVec2(Window::width() - 400, 50));
+	ImGui::SetNextWindowPos(ImVec2(static_cast<float>(Window::width()) - 400.0f, 50.0f));
 	ImGui::Begin("Instructions", reinterpret_cast<bool*>(1));
 	ImGui::Text("Hold middle mouse = orbit camera\nScroll = zoom in / out\nShift + A = new object menu\nLeft click = select object\nLeft ctrl + left click = select multiple objects\nF = set last selected object as orbit point\nG = select move tool\nS = select scale tool\nR = select rotate tool");
 	ImGui::End();
